fix: replace gen_mosaic script to use TiTiler instead of manual
