+++ backend/global-mosaic/pyproject.toml
@@ -6,10 +6,11 @@ readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "affine>=2.4.0",
+    "aiohttp>=3.12.15",
     "mercantile>=1.2.1",
     "numpy>=2.3.2",
     "pmtiles>=3.4.1",
     "psycopg>=3.2.9",
-    "rio-tiler>=7.8.1",
+    "rasterio>=1.4.3",
     "shapely>=2.1.1",
 ]
